refactor(lsp): unify command detection
